+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- 2019-06-17 : (PLB2B-412)  : Upgrade to Swift 5
+
 ### Changed
 ### Fixed
 ### Removed